Moved directory structure around to cater for data sync
